==> 7_survival_analysis_os_pfs.Rmd <==
---
title: "7_survival_analysis_os_pfs"
output: html_document
author: "Carlos Fernandez Moro and Natalie Geyer"
date: '`r Sys.Date()`'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Generated on: `r Sys.time()`

Import required packages
```{r}
library(tidyverse)
library(survival)
library(survminer)
library(corrplot)
library(prodlim)
```

```{r}
clinPatAnnotFn <- "./output/clin_data_annot.csv"
```

Import data
```{r}
clinPatAnnotData <- read.csv(clinPatAnnotFn)

clinPatAnnotData$gp_predo3 <- as.factor(clinPatAnnotData$gp_predo3)
clinPatAnnotData$gp_predo3 <- fct_relevel(clinPatAnnotData$gp_predo3, "R", "D", "P")

clinPatAnnotData$desmo_ks <- as.factor(clinPatAnnotData$desmo_ks)
clinPatAnnotData$desmo_ks <- fct_relevel(clinPatAnnotData$desmo_ks, "[0,33]", "(33,99.9]", "(99.9,100]")
```

Parametrized survival function (to avoid code repetition)
```{r}
plotSurv <- function(grouping, fun_data, surv_type, do_facet, cohort_type = "All") {
  
params <- list(grouping = substitute(grouping), fun_data = substitute(fun_data))

if(surv_type == "OS") {
  expr <- substitute(survfit(Surv(time = time_fu_os, event = status_fu_os,) ~ grouping, data = fun_data), params)
} else if(surv_type == "Liver PFS") {
  expr <- substitute(survfit(Surv(time = time_fu_prog, event = status_fu_prog,) ~ grouping, data = fun_data), params)
}
fit_object <- eval.parent(expr)
print(fit_object)

# All  
p <- ggsurvplot(fit_object, data = fun_data, 
    title = paste(surv_type, "- ", cohort_type),
    xlab = "Years", 
    ylab = "Survival probability",
    pval = TRUE, 
    risk.table = TRUE,
    risk.table.y.text = FALSE, 
    risk.table.height = 0.3, 
    palette = c("#d73027", "#313695",  "#fdae61"),
    xscale = 365.25, # converts days->years at xscale
    break.time.by=365.25) # sets the axis breaks to one year
print(p)

# Neoadjuvant vs Chemonaïve
if(do_facet) {
  p <- ggsurvplot_facet(fit_object, data = fun_data, 
      title = surv_type,
      xlab = "Years", 
      ylab = "Survival probability",
      pval = TRUE, 
      facet.by = "neoadjuvant", 
      palette = c("#d73027", "#313695",  "#fdae61"),
      xscale = 365.25, 
      break.time.by=365.25)  
  print(p)
 }
}
```

Estimation of the median follow-up time (reverse Kaplan-Meier)
```{r}
quantile(prodlim(Hist(time_fu_os, status_fu_os)~1,data=clinPatAnnotData,reverse=TRUE))
```

Survival plots
```{r}
# Global OS All
plotSurv(1, clinPatAnnotData, "OS", FALSE)
# Global OS Neoadjuvant
plotSurv(1, filter(clinPatAnnotData, neoadjuvant == "Yes"), "OS", FALSE, "Neoadjuvant")
# Global OS Chemonaïve
plotSurv(1, filter(clinPatAnnotData, neoadjuvant == "No"), "OS", FALSE, "Chemonaïve")
# ------------------------
# Global PFS All
plotSurv(1, clinPatAnnotData, "Liver PFS", FALSE)
# Global OS Neoadjuvant
plotSurv(1, filter(clinPatAnnotData, neoadjuvant == "Yes"), "Liver PFS", FALSE, "Neoadjuvant")
# Global OS Chemonaïve
plotSurv(1, filter(clinPatAnnotData, neoadjuvant == "No"), "Liver PFS", FALSE, "Chemonaïve")

# Neoadjuvant
plotSurv(neoadjuvant, clinPatAnnotData, "OS", FALSE)
plotSurv(neoadjuvant, clinPatAnnotData, "Liver PFS", FALSE)

# Predominant GP (D, R, P)
plotSurv(gp_predo3, clinPatAnnotData, "OS", TRUE)
plotSurv(gp_predo3, filter(clinPatAnnotData, neoadjuvant == "Yes"), "OS", FALSE, "Neoadjuvant")
plotSurv(gp_predo3, filter(clinPatAnnotData, neoadjuvant == "No"), "OS", FALSE, "Chemonaïve")
plotSurv(gp_predo3, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(gp_predo3, filter(clinPatAnnotData, neoadjuvant == "Yes"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(gp_predo3, filter(clinPatAnnotData, neoadjuvant == "No"), "Liver PFS", FALSE, "Chemonaïve")
```

Multivariate analysis (CCox proportional hazards model)
```{r, fig.width= 10}

# Fit a Cox proportional hazards model
# Alternatively, build model with continuous data instead of discretized

# UNI - OS - desmo_ks
fit.coxph <- coxph(Surv(time = time_fu_os, event = status_fu_os) ~ desmo_ks, data = clinPatAnnotData)
summary(fit.coxph)
ggforest(fit.coxph, data = clinPatAnnotData, main = "UNI - OS All - desmo_ks")

# MULTI - OS desmo_ks
fit.coxph <- coxph(Surv(time = time_fu_os, event = status_fu_os) ~ age_ + 
                     primary_side + 
                     num_tumors + 
                     max_tum_diameter + 
                     marginal +
                     tum_percent_annot_ + 
                     desmo_ks, 
                   data = clinPatAnnotData)
summary(fit.coxph)
ggforest(fit.coxph, data = clinPatAnnotData, main = "OS All - desmo_ks")

# UNI - Liver PFS - desmo_ks
fit.coxph <- coxph(Surv(time = time_fu_prog, event = status_fu_prog) ~ desmo_ks, data = clinPatAnnotData)
summary(fit.coxph)
ggforest(fit.coxph, data = clinPatAnnotData, main = "UNI - Liver PFS All - desmo_ks")

# MULTI - Liver PFS desmo_ks
fit.coxph <- coxph(Surv(time = time_fu_prog, event = status_fu_prog) ~ primary_side + 
                     sync_meta + 
                     num_tumors + 
                     max_tum_diameter + 
                     marginal + 
                     desmo_ks, data = clinPatAnnotData)
summary(fit.coxph)
ggforest(fit.coxph, data = clinPatAnnotData, main = "Liver PFS All - desmo_ks")
```
